requirements.txt
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
aiogram==3.4.1
aiogram-forms==1.1.1
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
aiosqlite==0.20.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pylint==3.1.0
SQLAlchemy==2.0.29
pillow==10.3.0
django-filter==24.2
